+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - WebGL runtime error for Unity2020 build due to unsupported compression setting.
 - Unity WebGL template path wasn't updating to Zappar2020 on `Zappar/Editor/UpdateProjectSettingsToPublish`.
+- Dialog pop-up when enabling SRP for Zappar package and Unity Universal Rendering Package is not imported in project.
 
 
 ## [1.1.0-preview.2] - 2021-07-05