+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0-preview.1] - 2021-XX-XX
+### Added
+- Zappar image training in editor
+
+### Fixed
+- 
+
+
 ## [1.1.1-preview.1] - 2021-07-14
 ### Added
 - Camera-based realtime reflection